requirements.txt
# For backend
Flask==3.1.1
Flask-CORS==6.0.1

# Streamlit frontend
streamlit==1.41.1

# Gemini AI integration
google-generativeai==0.8.5

# Gmail API with OAuth
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2

# Environment variables
python-dotenv==1.1.1

# HTTP requests (for frontend-backend communication)
requests==2.32.3